Move NetPeerUtils related tests to a new groovy class
